moved widget creation to a util
